requirements.txt
streamlit
google-genai
numpy
pypdf
python-dotenv
